backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
